Add confirmation dialog when deleting items

+++ main.js
@@ -96,8 +96,6 @@ function addItemToDOM(
   buttonsContainer.appendChild(editButton);
   buttonsContainer.appendChild(removeButton);
   li.appendChild(buttonsContainer);
-
-  // Add li to the DOM
   itemList.appendChild(li);
 }
 
@@ -198,9 +196,11 @@ function onItemEdit(item) {
 }
 
 function deleteItem(item) {
-  item.remove();
-  deleteFromLocalStorage(item.getAttribute("data-key"));
-  reloadList();
+  if (confirm("Remove item from resume?")) {
+    item.remove();
+    deleteFromLocalStorage(item.getAttribute("data-key"));
+    reloadList();
+  }
 }
 
 function deleteFromLocalStorage(index) {
@@ -210,12 +210,13 @@ function deleteFromLocalStorage(index) {
 }
 
 function clearList() {
-  while (itemList.firstChild) {
-    itemList.removeChild(itemList.firstChild);
+  if (confirm("Clear resume? You cannot undo this action.")) {
+    while (itemList.firstChild) {
+      itemList.removeChild(itemList.firstChild);
+    }
+    localStorage.removeItem("items");
+    reloadList();
   }
-
-  localStorage.removeItem("items");
-  reloadList();
 }
 
 function reloadList() {
